Extract waveform edit fade paint helpers
